Reducing npm package size by creating a separate Readme file and also, excluding test source code from the final npm package
