+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cabinets-react",
-  "version": "1.1.0",
+  "version": "1.1.5",
   "description": "Cabinets-react is the simplest Global State Management for React.js",
   "main": "src/module.js",
   "scripts": {
@@ -26,7 +26,7 @@
   },
   "homepage": "https://github.com/andreidim/cabinets-react.js#readme",
   "dependencies": {
-    "cabinets": "^1.0.5"
+    "cabinets": "^1.0.6"
   },
   "devDependencies": {
     "@babel/core": "^7.12.13",
