+++ test/CounterMessage.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { useStoreHook } from "../module.js";
+import { useStoreHook } from "../src/module.js";
 
 export default function CounterMessage() {
