+++ test/CounterStore.js
